Main_Gateway/backend/requirements.txt
fastapi>=0.115.12,<0.116.0
uvicorn[standard]>=0.34.1,<0.35.0
pydantic>=2.11.3,<3.0.0
python-dotenv>=1.1.0,<2.0.0
httpx>=0.28.1,<0.29.0
python-jose[cryptography]>=3.4.0,<4.0.0
requests>=2.32.3,<3.0.0
sqlalchemy>=2.0.40,<3.0.0
pyodbc>=5.2.0,<6.0.0
pandas>=2.2.3,<3.0.0
websockets>=15.0.1,<16.0.0
websocket-client>=1.8.0,<2.0.0
gunicorn>=23.0.0
azure-cosmos>=4.5.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.9.0
azure-monitor-opentelemetry>=1.6.0,<1.7.0
azure-core>=1.32.0
aiohttp>=3.8.0
psycopg2-binary>=2.9.0